backend/pyproject.toml
[tool.poetry]
name = "threadist-backend"
version = "0.1.0"
description = "Backend API for Threadist - Reddit story narration app"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "threadist_backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
supabase = "^1.0.3"
elevenlabs = "^0.2.26"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
redis = "^5.0.1"
celery = "^5.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 